CZU-KNOWLEDGE-SHARE-PROJECT/knowledgeShare/requirements.txt
asgiref==3.4.1
astroid==2.9.0
attrs==21.2.0
black==21.11b1
click==8.0.3
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.1
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt
djangorestframework-simplejwt
drf-spectacular==0.21.0
gunicorn==20.1.0
importlib-resources==5.4.0
inflection==0.5.1
isort==5.10.1
jsonschema==4.2.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.3
PyJWT==2.3.0
pylint==2.12.2
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing==3.7.4.3
typing-extensions==4.0.1
uritemplate==4.1.1
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.6.0
